requirements.txt
# Core Dependencies Only - Start with these
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Essential for Token Analysis
aiohttp==3.8.6
requests==2.31.0
web3==6.11.0

# Database
sqlalchemy[asyncio]==2.0.22
asyncpg==0.28.0
psycopg2-binary==2.9.9
alembic==1.12.0
redis==5.0.1

# Data Processing
pandas==2.1.1
numpy==1.24.3

# ML (Add when ready)
scikit-learn==1.3.1
joblib==1.3.2

# Utils
python-dotenv==1.0.0
loguru==0.7.2

# Celery
celery==5.3.1
flower==2.0.1

# Security & Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
